pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "doodle-doc"
version = "0.1.0"
description = "Sketch-based search for handwritten PDF notes"
requires-python = ">=3.11"
dependencies = [
    # API
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    # PDF & Image Processing
    "pymupdf>=1.23.0",
    "pillow>=10.0.0",
    "opencv-python-headless>=4.9.0",
    # ML
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "transformers>=4.37.0",
    "accelerate>=0.26.0",
    "sentencepiece>=0.1.99",
    "protobuf>=4.25.0",
    # Vector Search
    "faiss-cpu>=1.7.4",
    # Text Search
    "rank-bm25>=0.2.2",
    # Config & Database
    "pyyaml>=6.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "google-genai>=1.56.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[project.scripts]
doodle-doc = "doodle_doc.cli:main"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "httpx>=0.28.1",
    "reportlab>=4.4.7",
    "matplotlib>=3.10.8",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
